src/agent-with-local-tools/requirements.txt
azure-ai-agentserver-agentframework==1.0.0b12
python-dotenv
azure-identity
httpx

# OBO token validation and acquisition
msal>=1.24.0
PyJWT[crypto]>=2.8.0
requests>=2.31.0

# Azure Monitor / OpenTelemetry
azure-monitor-opentelemetry-exporter>=1.0.0b46
opentelemetry-sdk>=1.39.0
opentelemetry-api>=1.39.0
